requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# Azure
azure-data-tables
azure-functions
azure-functions-durable
azure-identity
azure-keyvault-secrets
azure-storage-blob # Use datalake when possible
azure-storage-file-datalake
azure-storage-queue

# Database
pyodbc
sqlalchemy
shillelagh

# Data Manipulation
geojson
pyarrow

# Flask
Flask
Flask-SQLAlchemy
Flask-Session
Flask-SQLAlchemy-Session
Flask-Restless-NG

# Identity/Authenticity/Authorization
py-abac
pyjwt
requests-auth-aws-sigv4
Authlib

# Maths
area
numpy
shapely

# Other
pyproj
requests